requirements.txt
Flask==1.0.3
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.4.0
marshmallow==2.19.5
psycopg2==2.8.3
argon2-cffi==19.1.0
passlib==1.7.1
